Split classes into one-class-per-file; drop section-banner comments
Per copilot-instructions.md, every class lives in its own .cs file and
the '// ---- Foo ----' banner comments are gone. Updated the agent
instructions to make these rules explicit.
