+++ src/ComposeNet.Compose/NavigationBar.cs
@@ -1,0 +1,36 @@
+using AndroidX.Compose.Runtime;
+
+namespace ComposeNet;
+
+/// <summary>
+/// Material 3 <c>NavigationBar</c>. Container for
+/// <see cref="NavigationBarItem"/> children laid out horizontally:
+/// <code>
+/// new NavigationBar
+/// {
+///     new NavigationBarItem(selected: tab == 0, onClick: () =&gt; tab.Value = 0)
+///     {
+///         Icon = new Text("🏠"), Label = new Text("Home"),
+///     },
+///     new NavigationBarItem(selected: tab == 1, onClick: () =&gt; tab.Value = 1)
+///     {
+///         Icon = new Text("⚙"), Label = new Text("Settings"),
+///     },
+/// }
+/// </code>
+/// </summary>
+public sealed class NavigationBar : ComposableContainer
+{
+    internal override void Render(IComposer composer)
+    {
+        // Capture the RowScope receiver (p0 of the Function3) and publish
+        // it so child NavigationBarItems can pass it to their underlying
+        // RowScope-extension static.
+        var content = new ComposableLambda3((scope, c) =>
+        {
+            using var _ = RenderContext.PushScope(scope);
+            RenderChildren(c);
+        });
+        ComposeBridges.NavigationBar(content, composer);
+    }
+}
