+++ src/ComposeNet.Compose/Surface.cs
@@ -1,0 +1,16 @@
+using AndroidX.Compose.Runtime;
+
+namespace ComposeNet;
+
+/// <summary>
+/// Material 3 non-interactive <c>Surface</c> — applies background color,
+/// elevation, and clipping to its content.
+/// </summary>
+public sealed class Surface : ComposableContainer
+{
+    internal override void Render(IComposer composer)
+    {
+        var content = new ComposableLambda2(c => RenderChildren(c));
+        ComposeBridges.Surface(content, composer);
+    }
+}
